Fixed stop container script

+++ scripts/stop_container.sh
@@ -3,6 +3,11 @@ set -e
 set -o pipefail
 # Stop the running container (if any)
 echo "removing container..."
-containerid=`sudo docker ps | awk -F" " '{print $1}' | tail -1`
-printf "Stopping %s" $containerid
-sudo docker rm -f $containerid
+containerid=`sudo docker ps -q | awk -F" " '{print $1}' | tail -1`
+printf "containerid set to: %s \n" $containerid 
+if [[ -z $containerid ]]; then
+    printf "No existing container: %s \n" $containerid
+else
+    printf "Stopping %s \n" $containerid
+    sudo docker rm -f $containerid
+fi
